refactor: URLParser, parseURL
