+++ SwiftDiffusion/Utilities/Parsers/ParseGenerationData/PromptView+ParseCivitai.swift
